add cast.rs file and moved all castings functions to it
