+++ src/main.rs
@@ -1,12 +1,10 @@
-use ark_ff::PrimeField as ArkPF;
+mod cast;
+
 use ark_std::str::FromStr;
-use babyjubjub_ark::{new_key, verify, Fq, Fr, Point, PrivateKey, Signature};
+use babyjubjub_ark::{new_key, verify, Fq, Point, PrivateKey};
 use ff_ce::PrimeField;
 use hex;
-use num::{BigInt, Num};
-use num_bigint::BigUint;
 use poseidon_rs::{Fr as FrPoseidon, Poseidon};
-use sha2::{Digest, Sha256};
 use std::env;
 use std::fs::{self, File};
 use std::io::{self, Read, Write};
@@ -157,8 +155,8 @@ fn main() {
                 print_u8_array(&private_key.key, "dec");
                 println!("");
 
-                println!("public key Field X: {}", fq_to_dec_string(&public_key.x));
-                println!("public key Field Y: {}", fq_to_dec_string(&public_key.y));
+                println!("public key Field X: {}", cast::fq_to_dec_string(&public_key.x));
+                println!("public key Field Y: {}", cast::fq_to_dec_string(&public_key.y));
             } else if output_format == "hex" {
                 // Print private key
                 print!("private key: ");
@@ -167,8 +165,8 @@ fn main() {
 
                 // Print public key
                 print!("public key: ");
-                let hex_string_x = fq_to_hex_string(&public_key.x);
-                let hex_string_y = fq_to_hex_string(&public_key.y);
+                let hex_string_x = cast::fq_to_hex_string(&public_key.x);
+                let hex_string_y = cast::fq_to_hex_string(&public_key.y);
 
                 println!("{}{}", hex_string_x, hex_string_y);
             } else {
@@ -200,21 +198,21 @@ fn main() {
             let hash_fq = calculate_hash_fq(&message_to_sign_string, &hash_algorithm);
 
             // Print the hash
-            println!("message Hash: {}", fq_to_dec_string(&hash_fq));
+            println!("message Hash: {}", cast::fq_to_dec_string(&hash_fq));
 
             // Sign the message
             let signature = private_key.sign(hash_fq).expect("Failed to sign message");
 
             if output_format == "detailed" {
                 // Print signature
-                println!("Signature: R.X: {}", fq_to_dec_string(&signature.r_b8.x));
-                println!("Signature: R.Y: {}", fq_to_dec_string(&signature.r_b8.y));
-                println!("Signature: S: {}", fr_to_dec_string(&signature.s));
+                println!("Signature: R.X: {}", cast::fq_to_dec_string(&signature.r_b8.x));
+                println!("Signature: R.Y: {}", cast::fq_to_dec_string(&signature.r_b8.y));
+                println!("Signature: S: {}", cast::fr_to_dec_string(&signature.s));
             } else if output_format == "hex" {
                 // change signature variables to hex
-                let signature_x_hex = fq_to_hex_string(&signature.r_b8.x);
-                let signature_y_hex = fq_to_hex_string(&signature.r_b8.y);
-                let signature_s_hex = fr_to_hex_string(&signature.s);
+                let signature_x_hex = cast::fq_to_hex_string(&signature.r_b8.x);
+                let signature_y_hex = cast::fq_to_hex_string(&signature.r_b8.y);
+                let signature_s_hex = cast::fr_to_hex_string(&signature.s);
 
                 println!(
                     "Signature: {}{}{}",
@@ -241,8 +239,8 @@ fn main() {
             let hash_fq = calculate_hash_fq(&message_to_verify_string, &hash_algorithm);
 
             // Create PublicKey and signature objects
-            let public_key = public_key_from_str(&public_key_hex_string).unwrap();
-            let signature = signature_from_str(&signature_string);
+            let public_key = cast::public_key_from_str(&public_key_hex_string).unwrap();
+            let signature = cast::signature_from_str(&signature_string);
 
             let correct = verify(public_key, signature, hash_fq);
 
@@ -277,8 +275,8 @@ fn save_private_key(filename: &str, private_key: &PrivateKey) -> io::Result<()> 
 fn save_public_key(filename: &str, public_key: Point) -> io::Result<()> {
     let mut file = File::create(filename)?;
 
-    let x_string_hex = fq_to_hex_string(&public_key.x);
-    let y_string_hex = fq_to_hex_string(&public_key.y);
+    let x_string_hex = cast::fq_to_hex_string(&public_key.x);
+    let y_string_hex = cast::fq_to_hex_string(&public_key.y);
 
     write!(file, "{}{}", x_string_hex, y_string_hex).unwrap();
 
@@ -304,41 +302,6 @@ fn load_private_key(filename: &str) -> io::Result<PrivateKey> {
     Ok(private_key)
 }
 
-fn public_key_from_str(key_string_hex: &str) -> io::Result<Point> {
-    // Split the hex string into x and y parts
-    let (x_string_hex, y_string_hex) = key_string_hex.split_at(64);
-
-    // Parse hex strings into BigUint
-    let x_decimal = BigUint::from_str_radix(x_string_hex, 16).unwrap();
-    let y_decimal = BigUint::from_str_radix(y_string_hex, 16).unwrap();
-
-    // Convert BigUint to Fq
-    let x = Fq::from(x_decimal);
-    let y = Fq::from(y_decimal);
-
-    Ok(Point { x, y })
-}
-
-fn signature_from_str(signature_string_hex: &str) -> Signature {
-    // Split the string at indices 64 and 128
-    let (x_string_hex, temp) = signature_string_hex.split_at(64);
-    let (y_string_hex, s_string_hex) = temp.split_at(64);
-
-    // Parse hex strings into BigUint
-    let x_decimal = BigUint::from_str_radix(x_string_hex, 16).unwrap();
-    let y_decimal = BigUint::from_str_radix(y_string_hex, 16).unwrap();
-    let s_decimal = BigUint::from_str_radix(s_string_hex, 16).unwrap();
-
-    // Convert BigUint to Fq
-    let x = Fq::from(x_decimal);
-    let y = Fq::from(y_decimal);
-    let s = Fr::from(s_decimal);
-
-    let r_b8 = Point { x, y };
-
-    Signature { r_b8, s }
-}
-
 fn print_u8_array(arr: &[u8], format: &str) {
     for &element in arr {
         if format == "hex" {
@@ -349,143 +312,28 @@ fn print_u8_array(arr: &[u8], format: &str) {
     }
 }
 
-pub fn hash_to_bigint(hash: &[u8]) -> BigUint {
-    // Reverse the bytes because BigUint uses little-endian order
-    let reversed_hash: Vec<u8> = hash.iter().rev().cloned().collect();
-
-    // Create a BigUint from the reversed hash bytes
-    BigUint::from_bytes_le(&reversed_hash)
-}
-
-pub fn fq_to_hex_string(num: &Fq) -> String {
-    // convert to a decimal string
-    let num_string = num.to_string();
-
-    // Parse the decimal string into a hex
-    let num_decimal = BigUint::parse_bytes(num_string.as_bytes(), 10).unwrap();
-    let num_hex_string = format!("{:0>64x}", num_decimal);
-
-    // return the hex string
-    num_hex_string
-}
-
-pub fn fq_to_dec_string(num: &Fq) -> String {
-    // convert to a decimal string
-    let num_string = num.to_string();
-
-    // Parse the decimal string into a hex
-    let num_decimal = BigUint::parse_bytes(num_string.as_bytes(), 10).unwrap();
-
-    // return the hex string
-    num_decimal.to_string()
-}
-
-pub fn fr_to_hex_string(num: &Fr) -> String {
-    // convert to a decimal string
-    let num_string = num.to_string();
-
-    // Parse the decimal string into a hex
-    let num_decimal = BigUint::parse_bytes(num_string.as_bytes(), 10).unwrap();
-    let num_hex_string = format!("{:0>64x}", num_decimal);
-
-    // return the hex string
-    num_hex_string
-}
-
-pub fn fr_to_dec_string(num: &Fr) -> String {
-    // convert to a decimal string
-    let num_string = num.to_string();
-
-    // Parse the decimal string into a hex
-    let num_decimal = BigUint::parse_bytes(num_string.as_bytes(), 10).unwrap();
-
-    // return the hex string
-    num_decimal.to_string()
-}
-
-pub fn hash256_as_string(message: &str) -> String {
-    // create hash of the message
-    let mut hasher = Sha256::new();
-    hasher.update(message.as_bytes());
-    let hashed_message = hasher.finalize();
-
-    //Convert the hash result to a BigInt<4> -> hex string -> fq
-    let hashed_message_bigint = hash_to_bigint(&hashed_message[..]);
-    let hashed_message_string = hashed_message_bigint.to_str_radix(10);
-
-    hashed_message_string
-}
-
-pub fn hash_as_hex_string(message: &String) -> String {
-    // create hash of the message
-    let mut hasher = Sha256::new();
-    hasher.update(message.as_bytes());
-    let hashed_message = hasher.finalize();
-
-    //Convert the hash result to a BigInt<4> -> hex string -> fq
-    let hashed_message_bigint = hash_to_bigint(&hashed_message[..]);
-    let hashed_hex_message_string = format!("{:0>64x}", hashed_message_bigint);
-
-    hashed_hex_message_string
-}
-
-fn convert_hex_to_dec(hex_str: &str) -> String {
-    let hex_value = if hex_str.starts_with("0x") {
-        &hex_str[2..] // Remove the '0x' prefix
-    } else {
-        hex_str
-    };
-
-    BigInt::from_str_radix(hex_value, 16)
-        .map(|dec_value| dec_value.to_string())
-        .unwrap_or_else(|_| String::from("Invalid hex number"))
-}
-
-/// Function for converting bytes to fields.
-fn bytes_to_fields(bs: &[u8]) -> Vec<Fq> {
-    let reversed_split_bytes: Vec<Vec<u8>> =
-        bs.iter()
-            .rev()
-            .fold(vec![vec![]], |mut acc: Vec<Vec<u8>>, b| {
-                let n = acc.len();
-                if acc[n - 1].len() == PACKED_BYTE_LEN {
-                    acc.push(vec![*b]);
-                    acc
-                } else {
-                    acc[n - 1].push(*b);
-                    acc
-                }
-            });
-    let packed_fields = reversed_split_bytes
-        .iter()
-        .map(|bs| bs.iter().rev().copied().collect::<Vec<u8>>())
-        .map(|bs| Fq::from_be_bytes_mod_order(&bs))
-        .rev()
-        .collect::<Vec<Fq>>();
-    packed_fields
-}
 
 #[test]
 fn byte_packing_test() {
     let bytes: &[u8] = &[0x01, 0x02, 0x03];
-    let fields = bytes_to_fields(bytes);
+    let fields = cast::bytes_to_fields(bytes);
     assert!(fields.len() == 1);
-    assert!(fields[0] == Fq::from_str(&convert_hex_to_dec("0x010203")).unwrap());
+    assert!(fields[0] == Fq::from_str(&cast::hex_to_dec("0x010203")).unwrap());
 
     let bytes: Vec<u8> = (0..64).collect();
-    let fields = bytes_to_fields(&bytes);
+    let fields = cast::bytes_to_fields(&bytes);
     assert!(fields.len() == 3);
-    assert!(fields[0] == Fq::from_str(&convert_hex_to_dec("0x0001")).unwrap());
+    assert!(fields[0] == Fq::from_str(&cast::hex_to_dec("0x0001")).unwrap());
     assert!(
         fields[1]
-            == Fq::from_str(&convert_hex_to_dec(
+            == Fq::from_str(&cast::hex_to_dec(
                 "0x02030405060708090a0b0c0d0e0f101112131415161718191a1b1c1d1e1f20"
             ))
             .unwrap()
     );
     assert!(
         fields[2]
-            == Fq::from_str(&convert_hex_to_dec(
+            == Fq::from_str(&cast::hex_to_dec(
                 "0x2122232425262728292a2b2c2d2e2f303132333435363738393a3b3c3d3e3f"
             ))
             .unwrap()
@@ -500,7 +348,7 @@ fn calculate_hash_fq(message_to_verify_string: &str, hash_algorithm: &str) -> Fq
         let bytes = message_to_verify_string.as_bytes();
 
         // Pack the message bytes into right-aligned 31-byte chunks
-        let fr_vector: Vec<FrPoseidon> = bytes_to_fields(bytes)
+        let fr_vector: Vec<FrPoseidon> = cast::bytes_to_fields(bytes)
             .iter()
             .map(|&b| FrPoseidon::from_str(&b.to_string()).unwrap())
             .collect();
@@ -513,13 +361,13 @@ fn calculate_hash_fq(message_to_verify_string: &str, hash_algorithm: &str) -> Fq
 
         // turn into a string
         let mut poseidon_hash_str = poseidon_hash.into_repr().to_string();
-        poseidon_hash_str = convert_hex_to_dec(&poseidon_hash_str);
+        poseidon_hash_str = cast::hex_to_dec(&poseidon_hash_str);
 
         // turn the hash into Fq
         hash_fq = Fq::from_str(&poseidon_hash_str).unwrap();
     } else if hash_algorithm == "sha256" {
         //  hash the message
-        let hashed_sha256_message_string = hash256_as_string(message_to_verify_string);
+        let hashed_sha256_message_string = cast::hash256_as_string(message_to_verify_string);
 
         // turn the hash into Fq
         hash_fq = Fq::from_str(&hashed_sha256_message_string).unwrap();
@@ -535,7 +383,7 @@ fn test_poseidon_hash() {
     let msg = "This is a run-through of the Poseidon permutation function.";
     let hash = calculate_hash_fq(msg, "poseidon");
     assert!(
-        hash == Fq::from_str(&convert_hex_to_dec(
+        hash == Fq::from_str(&cast::hex_to_dec(
             "0x0b5de89054f5ff651f919eb397f4a125e9ba2aebd175dd809fe8fd02569d8087"
         ))
         .unwrap()
@@ -573,3 +421,4 @@ pub fn bad_command(command: &str) {
 fn file_exists(file_path: &str) -> bool {
     fs::metadata(file_path).is_ok()
 }
+
